added messaging redirect to the most recent mesage
